Refactor transcribe function
I added a conditional flag to the main transcribe function that allows us to enable or disable the simple, rule based post processing of the raw transcript.

This way, when we add a more complex post processing solution, we can use that instead with a simple change to that parameter.