Refactor(features.py)
This commit changes the code for the NLP Feature extractor class. The init method was modified to no longer take a path for the output location of the features.

The logic is that the output_path variable in create_features handles this and there is no need to duplicate this.