Add model iterations:
This commit explores the addition of two new model iterations for the keras recurrent models.

Two new models featuring only one recurrent layer of each type (SimpleRNN and LSTM) were added to the Model-Iteration.ipynb jupyter notebook.

The models were also saved to the keras-models/ directory to allow for reproductible results later down the line.

The performance of the models was also documented inside the model iteration log file.

The error analysis text for the keras models was also rewritten, accounting for the two new models. It was improved to better document the results and emphasize how the models performed better with two recurrent layers compared to one.